api/requirements.txt
pdfplumber==0.10.3
google-generativeai==0.7.2
flask>=2.3.3
Werkzeug>=2.3.7
tenacity==8.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0